requirements.txt
ariadne>=0.13.0,<0.13.99
django>=2.2
python-dateutil==2.8.1
